Create Basic ToDo List
Create a list to store your to-do items, and initialize it with a few sample items. Create a function called "add_item" that takes in a string argument representing a new to-do item, and appends it to the list. Create a function called "remove_item" that takes in a string argument representing a to-do item, and removes it from the list. Create a function called "view_list" that prints out all the items in the list. Create a function called "check_item" that takes in a string argument representing a to-do item, and returns a boolean indicating whether the item is in the list or not. Create a function called "clear_list" that removes all items from the list. Create a function called "filter_list" that takes in a string argument and returns a new list containing only the items that contain the string. This function will use list comprehension. Create a variable called "running" and initialize it to True. This variable will control the main loop of the program. Use a while loop that continues as long as "running" is True. Inside the loop, print out a menu of options for the user (add item, remove item, view list, check item, clear list, filter list, and quit). Ask the user to select an option. Use an if-elif-else block to handle the user's selection. For example, if the user selects "add item", call the "add_item" function and pass in the item they want to add. If the user selects "remove item", call the "remove_item" function and pass in the item they want to remove. If the user selects "view list", call the "view_list" function. In the filter list option, it will call the filter_list function and pass the string which the user wants to filter the list on. And so on. In the "quit" option of the menu, set the "running" variable to False to exit the loop and end the program.

This project allows for the user to add, remove, and clear items from the list. They can also check if an item is on the list, filter the list based on a string, and view the entire list. This program uses loops, conditionals, and list comprehension in a beginner level and is around 130 lines.

Step-by-Step process:

Initialize an empty list to store the to-do items.

Define the "add_item" function, which takes in a string argument and appends it to the list.

Define the "remove_item" function, which takes in a string argument and removes the first occurrence of that string from the list.

Define the "view_list" function, which prints out all the items in the list.

Define the "check_item" function, which takes in a string argument and returns a boolean indicating whether the item is in the list or not.

Define the "clear_list" function, which removes all items from the list.

Define the "filter_list" function, which takes in a string argument and returns a new list containing only the items that contain the string using list comprehension.

Create a variable called "running" and initialize it to True.

Use a while loop that continues as long as "running" is True.

Inside the loop,Create a list to store your to-do items, and initialize it with a few sample items.

Create a function called "add_item" that takes in a string argument representing a new to-do item, and appends it to the list.

Create a function called "remove_item" that takes in a string argument representing a to-do item, and removes it from the list.

Create a function called "view_list" that prints out all the items in the list.

Create a function called "check_item" that takes in a string argument representing a to-do item, and returns a boolean indicating whether the item is in the list or not.

Create a function called "clear_list" that removes all items from the list.

Create a function called "list_comprehension" that takes in a keyword and returns a new list containing only the items from the to-do list that contain that keyword.

Create a variable called "running" and initialize it to True. This variable will control the main loop of the program.

Use a while loop that continues as long as "running" is True. Inside the loop, print out a menu of options for the user (add item, remove item, view list, check item, clear list, list comprehension, and quit). Ask the user to select an option.

Use an if-elif-else block to handle the user's selection. For example, if the user selects "add item", call the "add_item" function and pass in the item they want to add. If the user selects "remove item", call the "remove_item" function and pass in the item they want to remove. If the user selects "view list", call the "view_list" function. If the user selects "list comprehension", call the "list_comprehension" function and pass in the keyword they want to filter by. And so on.

In the "quit" option of the menu, set the "running" variable to False to exit the loop and end the program.

This project incorporates the following topics:

Lists: to store the to-do items

Loops: to control the main loop of the program and to iterate through the list of to-do items

Conditionals: to handle the user's menu selection and to check if an item is in the list

Functions: to encapsulate the functionality of adding, removing, viewing, and checking items in the list

Strings: to store the to-do items and to take in user input

Dictionaries: to store the different options available to the user and their corresponding functions.